# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Akrsf swagger documentation.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Kong Chen",
      "url" : "http://kongch.com",
      "email" : "kongchen@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/cselp/rest",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/admin/ldap/users" : {
      "get" : {
        "summary" : "Ldap users",
        "description" : "",
        "operationId" : "getLdapUsers",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/admin/person/update" : {
      "get" : {
        "summary" : "updated persons",
        "description" : "",
        "operationId" : "getPersonUpdates",
        "parameters" : [ {
          "name" : "tabNo",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/admin/persons" : {
      "get" : {
        "summary" : "persons",
        "description" : "",
        "operationId" : "findPersons",
        "parameters" : [ {
          "name" : "squadronId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/api/report/flight" : {
      "get" : {
        "summary" : "Flight Report",
        "description" : "",
        "operationId" : "getFlightReport",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "to",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FlightSearchResult"
            }
          }
        }
      }
    },
    "/api/report/flights" : {
      "get" : {
        "summary" : "Flight Report",
        "description" : "",
        "operationId" : "findReportFlights",
        "parameters" : [ {
          "name" : "dep_from",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dep_to",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "arr_a",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dep_a",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "high",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "med",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "low",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "first",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "amount",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "order",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FlightSearchResult"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Flight" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dep_datetime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "arr_datetime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dep_airport_tzo" : {
          "type" : "integer",
          "format" : "int32"
        },
        "arr_airport_tzo" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dep_airport" : {
          "type" : "string"
        },
        "arr_airport" : {
          "type" : "string"
        },
        "flight" : {
          "type" : "string"
        },
        "plane" : {
          "type" : "string"
        },
        "tail" : {
          "type" : "string"
        },
        "stat_events" : {
          "$ref" : "#/definitions/StatEventsQty"
        },
        "stat_score" : {
          "type" : "integer",
          "format" : "int32"
        },
        "crew" : {
          "$ref" : "#/definitions/FlightCrew"
        },
        "events" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FlightEvent"
          }
        }
      }
    },
    "FlightCrew" : {
      "type" : "object",
      "properties" : {
        "task_num" : {
          "type" : "string"
        },
        "task_datetime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "type" : {
          "type" : "string"
        },
        "members" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FlightCrewMember"
          }
        }
      }
    },
    "FlightCrewMember" : {
      "type" : "object",
      "properties" : {
        "tab_num" : {
          "type" : "string"
        },
        "squadron" : {
          "type" : "string"
        },
        "division" : {
          "type" : "string"
        },
        "flight_role" : {
          "type" : "string"
        },
        "first" : {
          "type" : "string"
        },
        "last" : {
          "type" : "string"
        }
      }
    },
    "FlightEvent" : {
      "type" : "object",
      "properties" : {
        "time" : {
          "type" : "integer",
          "format" : "int64"
        },
        "phase" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "severity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "score" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "FlightSearchResult" : {
      "type" : "object",
      "properties" : {
        "found" : {
          "type" : "integer",
          "format" : "int32"
        },
        "total" : {
          "type" : "integer",
          "format" : "int64"
        },
        "first" : {
          "type" : "integer",
          "format" : "int32"
        },
        "flights" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Flight"
          }
        }
      }
    },
    "StatEventsQty" : {
      "type" : "object",
      "properties" : {
        "high" : {
          "type" : "integer",
          "format" : "int32"
        },
        "med" : {
          "type" : "integer",
          "format" : "int32"
        },
        "low" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    }
  }
}